Fix support for newer gpgme python bindings
The API for the python-gpg package changed drastically between Ubuntu
16.04 and 20.04. This commit updates system-image-server to the new API.

This required an API break inside gpg.py. It is no longer valid to
provide a string value for the key expiry, such as "2y". Now the gpg
library only takes the key *validity* as an integer number of seconds
past the generation time.