Use rsa4096 for archive-master and image-master
When new system-image-server users generate their keys, we want them to
be reasonably secure. For that reason, we should increase the key length
for archive-master and image-master, the long-lived keys of the keyring.

The short-lived keys are used more often and verified more often on slow
devices. To balance performance with security, keep the short-lived keys
at rsa2048.