Fix comparison of None against int
The line 'sorted(target_images - device_images)' threw an error
previously:

TypeError: '<' not supported between instances of 'int' and 'NoneType'

This error is caused by the sets containing None, which is the fallback
return when the image has no "base" key. In Python 2, an
"is-greater-than" comparison between int and None was valid, with None
always sorting below an int. In Python 3, this changed. Now we need to
find an int value which is below any possible input for the sort. -1
should do. This allows the sort to continue, and places the images
with no delta below the images with a delta in the final result.